backend/pyproject.toml
[tool.poetry]
name = "suas-backend"
version = "0.1.0"
description = "SUAS — Shut Up and Serve backend API"
authors = ["Fernando <fmacademiajr@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
google-cloud-firestore = "^2.16.0"
google-cloud-storage = "^2.17.0"
google-cloud-logging = "^3.10.0"
google-cloud-secret-manager = "^2.20.0"
anthropic = "^0.27.0"
google-generativeai = "^0.7.0"
apscheduler = "^3.10.4"
feedparser = "^6.0.11"
httpx = "^0.27.0"
pytrends = "^4.9.2"
simhash = "^2.1.2"
firebase-admin = "^6.5.0"
pytz = "^2024.1"
scikit-learn = "^1.5.0"
joblib = "^1.4.0"
pillow = "^10.3.0"
playwright = "^1.44.0"
python-multipart = "^0.0.9"
email-validator = "^2.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
respx = "^0.21.0"
ruff = "^0.4.0"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
env = [
    "APP_ENV=test",
    "GCP_PROJECT_ID=suas-local",
    "FIRESTORE_EMULATOR_HOST=localhost:8080",
    "ALLOWED_USER_EMAIL=test@example.com",
    "ANTHROPIC_API_KEY=test-key",
    "GEMINI_API_KEY=test-key",
    "FACEBOOK_PAGE_ID=test-page-id",
    "FACEBOOK_PAGE_ACCESS_TOKEN=test-token",
    "REDDIT_CLIENT_ID=test-id",
    "REDDIT_CLIENT_SECRET=test-secret",
    "GCS_BUCKET_NAME=suas-test-bucket",
]

[tool.mypy]
python_version = "3.11"
strict = false
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
